saving articles

+++ src/main/java/pl/kololos/api/repositories/ArticlesRepository.java
@@ -3,5 +3,5 @@ package pl.kololos.api.repositories;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import pl.kololos.api.models.admin.Article;
 
-public interface ArticlesRepository extends PagingAndSortingRepository<Article, Integer> {
+public interface ArticlesRepository extends PagingAndSortingRepository<Article, Long> {
 }
